apptrix_web/requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
asgiref==3.4.1
async-timeout==4.0.1
asyncio==3.4.3
beautifulsoup4==4.10.0
bs4==0.0.1
charset-normalizer==2.0.9
defusedxml==0.7.1
Django==3.2.7
django-filter==21.1
djangorestframework==3.13.1
frozenlist==1.2.0
httpie==2.6.0
multidict==5.2.0
Pillow==9.0.0
Pygments==2.10.0
PySocks==1.7.1
pytz==2021.1
requests-toolbelt==0.9.1
soupsieve==2.3.1
sqlparse==0.4.2
tqdm==4.62.3
typing-extensions==4.0.1
utils==1.0.1
websocket-client==1.2.3
wget==3.2
yarl==1.7.2
